requirements.txt
# Core Data Analysis
pandas>=2.3.3
numpy>=2.3.5

# Web Scraping
google-play-scraper>=1.2.7
requests>=2.32.5
beautifulsoup4>=4.14.2

# NLP & Sentiment Analysis
transformers>=4.30.0
torch>=2.9.1
nltk>=3.9.2
langdetect>=1.0.9

# Database
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1

# Visualization
matplotlib>=3.10.7
seaborn>=0.13.2
wordcloud>=1.9.0

# Development (minimal Jupyter)
jupyter-client>=8.6.0
jupyter-core>=5.7.0
ipykernel>=6.0.0
